requirements.txt
# Core dependencies
fastapi
uvicorn
pydantic
requests
beautifulsoup4
pyngrok
nest-asyncio    
python-dotenv
streamlit

# LangChain and LangGraph dependencies
langchain
langchain-core
langchain-openai
langgraph

# OpenAI API
openai



#remove the versions from all above dependencies
